app/requirements.txt
boto3
PyPDF2
python-docx
